chore: extract shared package (packages/shared) with types and constants
- Create @drafto/shared package with Database types, API row types, and
  shared constants (MAX_FILE_SIZE, BUCKET_NAME, etc.)
- Web app's database.types.ts now re-exports from @drafto/shared
- Replace hardcoded constants across web app with shared imports
- All 513 unit/integration tests + 46 E2E tests passing
