+++ apps/web/src/lib/supabase/database.types.ts
@@ -1,177 +1,1 @@
-export type Json = string | number | boolean | null | { [key: string]: Json | undefined } | Json[];
-
-export interface Database {
-  public: {
-    Tables: {
-      profiles: {
-        Row: {
-          id: string;
-          is_approved: boolean;
-          is_admin: boolean;
-          display_name: string | null;
-          created_at: string;
-          updated_at: string;
-        };
-        Insert: {
-          id: string;
-          is_approved?: boolean;
-          is_admin?: boolean;
-          display_name?: string | null;
-          created_at?: string;
-          updated_at?: string;
-        };
-        Update: {
-          id?: string;
-          is_approved?: boolean;
-          is_admin?: boolean;
-          display_name?: string | null;
-          created_at?: string;
-          updated_at?: string;
-        };
-        Relationships: [
-          {
-            foreignKeyName: "profiles_id_fkey";
-            columns: ["id"];
-            isOneToOne: true;
-            referencedRelation: "users";
-            referencedColumns: ["id"];
-          },
-        ];
-      };
-      notebooks: {
-        Row: {
-          id: string;
-          user_id: string;
-          name: string;
-          created_at: string;
-          updated_at: string;
-        };
-        Insert: {
-          id?: string;
-          user_id: string;
-          name: string;
-          created_at?: string;
-          updated_at?: string;
-        };
-        Update: {
-          id?: string;
-          user_id?: string;
-          name?: string;
-          created_at?: string;
-          updated_at?: string;
-        };
-        Relationships: [
-          {
-            foreignKeyName: "notebooks_user_id_fkey";
-            columns: ["user_id"];
-            isOneToOne: false;
-            referencedRelation: "users";
-            referencedColumns: ["id"];
-          },
-        ];
-      };
-      notes: {
-        Row: {
-          id: string;
-          notebook_id: string;
-          user_id: string;
-          title: string;
-          content: Json | null;
-          is_trashed: boolean;
-          trashed_at: string | null;
-          created_at: string;
-          updated_at: string;
-        };
-        Insert: {
-          id?: string;
-          notebook_id: string;
-          user_id: string;
-          title?: string;
-          content?: Json | null;
-          is_trashed?: boolean;
-          trashed_at?: string | null;
-          created_at?: string;
-          updated_at?: string;
-        };
-        Update: {
-          id?: string;
-          notebook_id?: string;
-          user_id?: string;
-          title?: string;
-          content?: Json | null;
-          is_trashed?: boolean;
-          trashed_at?: string | null;
-          created_at?: string;
-          updated_at?: string;
-        };
-        Relationships: [
-          {
-            foreignKeyName: "notes_notebook_id_fkey";
-            columns: ["notebook_id"];
-            isOneToOne: false;
-            referencedRelation: "notebooks";
-            referencedColumns: ["id"];
-          },
-          {
-            foreignKeyName: "notes_user_id_fkey";
-            columns: ["user_id"];
-            isOneToOne: false;
-            referencedRelation: "users";
-            referencedColumns: ["id"];
-          },
-        ];
-      };
-      attachments: {
-        Row: {
-          id: string;
-          note_id: string;
-          user_id: string;
-          file_name: string;
-          file_path: string;
-          file_size: number;
-          mime_type: string;
-          created_at: string;
-        };
-        Insert: {
-          id?: string;
-          note_id: string;
-          user_id: string;
-          file_name: string;
-          file_path: string;
-          file_size: number;
-          mime_type: string;
-          created_at?: string;
-        };
-        Update: {
-          id?: string;
-          note_id?: string;
-          user_id?: string;
-          file_name?: string;
-          file_path?: string;
-          file_size?: number;
-          mime_type?: string;
-          created_at?: string;
-        };
-        Relationships: [
-          {
-            foreignKeyName: "attachments_note_id_fkey";
-            columns: ["note_id"];
-            isOneToOne: false;
-            referencedRelation: "notes";
-            referencedColumns: ["id"];
-          },
-          {
-            foreignKeyName: "attachments_user_id_fkey";
-            columns: ["user_id"];
-            isOneToOne: false;
-            referencedRelation: "users";
-            referencedColumns: ["id"];
-          },
-        ];
-      };
-    };
-    Views: Record<string, never>;
-    Functions: Record<string, never>;
-    Enums: Record<string, never>;
-  };
-}
+export type { Database, Json } from "@drafto/shared";
