+++ packages/shared/src/constants.ts
@@ -1,0 +1,7 @@
+export const MAX_TITLE_LENGTH = 255;
+export const MAX_NOTEBOOK_NAME_LENGTH = 100;
+export const MAX_FILE_SIZE = 26214400; // 25MB in bytes
+export const MAX_FILE_NAME_LENGTH = 255;
+export const DEBOUNCE_MS = 500;
+export const BUCKET_NAME = "attachments";
+export const SIGNED_URL_EXPIRY_SECONDS = 604800; // 7 days
